Completely redesign how the storage works
Instead of storing the sizes in the metadata, we generate IDs with all
necessary information to generate URLs. This has a benefit that metadata
is much smaller. Also, we don't need to resort to the ugly plugin hack
that we did before.

This change is obviously backwards incompatible, meaning that users will
have to start uploading anew after this version. But it's much better to
do it early on now, instead of later.

There are several backwards incomptabile changes. First is that there is
no more "Large 1600" and "Large 2048" size, since those cannot be
generated.

Second is that `UploadedFile#url` now without arguments return the link
to the photo page, instead of the "Original" source URL. For "Original"
source URL you know have to pass the string to `:size`.

Third is that `UploadedFile#flickr_url` has been removed, in favour of
the above solution.

Fourth is that there is no more `UploadedFile#width` and
`UploadedFile#height`, since we don't have that information anymore.
However, these shouldn't be needed, because by Flickr sizes have exactly
specified widths, and that's in most cases the only dimension that we
need. In landscape photos that's the height.